Analysis Toolkit
New analysis tools object and menu item
Use for ad-hoc analyses of Logs - could go in separate program, simpler
to keep here. Output to go to Reports

One tool included, which analyses species logs to produce a range
chart/phylogram for all species (culled to remove small species that
went extinct without issue). It currently assumed data starts at time 0
with species 1.

It's not as efficient as it might be - works fine on small runs, might
crash or just take an impractical amount of time to run for 'real' runs
with e.g. a million years of data, and/or lots and lots of species. fix
will be to set it to only look at some of the lines of data, not yet
implemented

Also needs a method of looking at genome change over life of a species -
data recorded in log, assembled into data structure, so shouldn't be too
much work.