New species ID system
Function Groups_2017 in analyser, plus modifications throughout.

This is a total analyser - all critters alive at the analyse point are
included, and all are grouped. No lazarus species are possible. Speed is
good for many small species, less good for a few large ones. Species IDs
now propogate from parents. When a species splits, the most diverse one
is retained as the original species, and all less diverse ones are given
new IDs.

Also removed UI settings relevant only to old system, and added UI
options to restrict breeding to species. Without this it is possible
that some species may breed with other species if they are genetically
close, in which case their species tag is randomly selected from one of
the parents.

It's also possible to turn off the genetic difference check for
breeding, though this does not work well at present and causes the
species ID algorithm to run very slowly.